move to schema syntax
