+++ api/src/schema.js
@@ -1,0 +1,15 @@
+import { buildSchema } from 'graphql';
+
+export default buildSchema(
+  `
+type Secret {
+  name: String
+  value: String
+}
+type Query {
+  server: String
+  secrets: [Secret]
+  token: String
+}
+`,
+);
