+++ redeploy.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+docker-compose -f docker-compose-app.yml build && \
+  docker-compose -f docker-compose-app.yml push && \
+  docker stack deploy --compose-file=docker-compose-app.yml app
